Gallery Block did no work with Theme Test Data
In the video tutorial Morten suggests to import the Theme Test Data (or Theme Unit Test, as the where could in the past). But when i do import them, the gallery blocks did not work as they used to do. All images where arranged in on vertical line instead of a grid. I could make it work by adding 5 characters. I do not know if that might break former versions of the gallery block.